Move class 'Guard' to 'utils' package
