+++ Kitodo/src/main/java/org/kitodo/production/helper/tasks/EmptyTask.java
@@ -19,8 +19,8 @@ import java.util.concurrent.ExecutionException;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.kitodo.exceptions.Guard;
 import org.kitodo.production.helper.Helper;
+import org.kitodo.utils.Guard;
 
 /**
  * The class EmptyTask is the base class for worker threads that operate
