+++ Kitodo/src/main/java/org/kitodo/production/interfaces/activemq/ActiveMQProcessor.java
@@ -27,12 +27,12 @@ import org.kitodo.config.enums.ParameterCore;
 import org.kitodo.data.database.beans.Client;
 import org.kitodo.data.database.beans.User;
 import org.kitodo.data.database.exceptions.DAOException;
-import org.kitodo.exceptions.Guard;
 import org.kitodo.exceptions.ProcessorException;
 import org.kitodo.production.enums.ReportLevel;
 import org.kitodo.production.helper.Helper;
 import org.kitodo.production.security.SecurityUserDetails;
 import org.kitodo.production.services.ServiceManager;
+import org.kitodo.utils.Guard;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
